requirements.txt
beautifulsoup4==4.14.3
langchain==1.1.3
langchain_core==1.2.0
langchain_openai==1.1.3
langfuse==3.10.6
langgraph==1.0.5
pandas==2.3.3
python-dotenv==1.2.1
qdrant_client==1.16.2
streamlit==1.51.0
typing_extensions==4.15.0
tiktoken